Move result and output functions out of main
